requirements.txt
python-dotenv==1.0.1
pydantic>=2.0.0
neo4j==5.20.0
langchain>=0.2.0
langchain-google-genai>=1.0.3
langchain-huggingface>=0.0.3
sentence-transformers>=2.7.0
chromadb>=0.5.0
fastapi==0.111.0
uvicorn==0.29.0
streamlit==1.35.0
streamlit-agraph==0.0.45
requests==2.31.0